services/inference-gateway-svc/requirements.txt
# AIVO Inference Gateway - S2-01 Requirements
# Multi-provider AI inference with PII scrubbing and intelligent routing

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0

# HTTP client
httpx==0.25.2

# Data validation and serialization
pydantic==2.5.0

# OpenTelemetry tracing
opentelemetry-api==1.21.0
opentelemetry-sdk==1.21.0
opentelemetry-instrumentation-fastapi==0.42b0
opentelemetry-instrumentation-httpx==0.42b0
opentelemetry-exporter-otlp==1.21.0

# Optional dependencies for specific providers
# Vertex AI (if ENABLE_VERTEX=true)
google-cloud-aiplatform==1.38.0

# AWS Bedrock (if ENABLE_BEDROCK=true) 
boto3==1.34.0
botocore==1.34.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2  # For testing HTTP clients

# Optional: for similarity computation in embeddings
# numpy==1.24.3  # Uncomment if using similarity endpoint
